+++ lib/tools/extreme-search.ts
